reminder jobs

+++ src/main/java/com/twistlet/falcon/model/service/ReminderService.java
@@ -6,7 +6,7 @@ import com.twistlet.falcon.model.entity.FalconAppointment;
 
 public interface ReminderService {
 
-	List<FalconAppointment> listAppointmentsNeedingReminders(int seconds);
+	List<FalconAppointment> listAppointmentsNeedingReminders(long seconds);
 
 	void sendNotification(FalconAppointment falconAppointment);
 }
